+++ src/js/methods.js
@@ -1,13 +1,11 @@
-  $.extend(prototype, {
-
     // Show the crop box manually
     crop: function () {
-      if (!this.built || this.disabled) {
+      if (!this.isBuilt || this.isDisabled) {
         return;
       }
 
-      if (!this.cropped) {
-        this.cropped = true;
+      if (!this.isCropped) {
+        this.isCropped = true;
         this.limitCropBox(true, true);
 
         if (this.options.modal) {
@@ -22,26 +20,24 @@
 
     // Reset the image and crop box to their initial states
     reset: function () {
-      if (!this.built || this.disabled) {
+      if (!this.isBuilt || this.isDisabled) {
         return;
       }
 
       this.image = $.extend({}, this.initialImage);
       this.canvas = $.extend({}, this.initialCanvas);
-
-      // Required for strict mode
       this.cropBox = $.extend({}, this.initialCropBox);
 
       this.renderCanvas();
 
-      if (this.cropped) {
+      if (this.isCropped) {
         this.renderCropBox();
       }
     },
 
     // Clear the crop box
     clear: function () {
-      if (!this.cropped || this.disabled) {
+      if (!this.isCropped || this.isDisabled) {
         return;
       }
 
@@ -52,7 +48,7 @@
         height: 0
       });
 
-      this.cropped = false;
+      this.isCropped = false;
       this.renderCropBox();
 
       this.limitCanvas(true, true);
@@ -68,32 +64,45 @@
      * Replace the image's src and rebuild the cropper
      *
      * @param {String} url
+     * @param {Boolean} onlyColorChanged (optional)
      */
-    replace: function (url) {
-      if (!this.disabled && url) {
+    replace: function (url, onlyColorChanged) {
+      if (!this.isDisabled && url) {
         if (this.isImg) {
-          this.replaced = true;
           this.$element.attr('src', url);
         }
 
-        // Clear previous data
-        this.options.data = null;
-        this.load(url);
+        if (onlyColorChanged) {
+          this.url = url;
+          this.$clone.attr('src', url);
+
+          if (this.isBuilt) {
+            this.$preview.find('img').add(this.$clone2).attr('src', url);
+          }
+        } else {
+          if (this.isImg) {
+            this.isReplaced = true;
+          }
+
+          // Clear previous data
+          this.options.data = null;
+          this.load(url);
+        }
       }
     },
 
     // Enable (unfreeze) the cropper
     enable: function () {
-      if (this.built) {
-        this.disabled = false;
+      if (this.isBuilt) {
+        this.isDisabled = false;
         this.$cropper.removeClass(CLASS_DISABLED);
       }
     },
 
     // Disable (freeze) the cropper
     disable: function () {
-      if (this.built) {
-        this.disabled = true;
+      if (this.isBuilt) {
+        this.isDisabled = true;
         this.$cropper.addClass(CLASS_DISABLED);
       }
     },
@@ -102,8 +111,8 @@
     destroy: function () {
       var $this = this.$element;
 
-      if (this.ready) {
-        if (this.isImg && this.replaced) {
+      if (this.isLoaded) {
+        if (this.isImg && this.isReplaced) {
           $this.attr('src', this.originalUrl);
         }
 
@@ -143,7 +152,7 @@
      */
     moveTo: function (x, y) {
       var canvas = this.canvas;
-      var changed = false;
+      var isChanged = false;
 
       // If "y" is not present, its default value is "x"
       if (isUndefined(y)) {
@@ -153,18 +162,18 @@
       x = num(x);
       y = num(y);
 
-      if (this.built && !this.disabled && this.options.movable) {
+      if (this.isBuilt && !this.isDisabled && this.options.movable) {
         if (isNumber(x)) {
           canvas.left = x;
-          changed = true;
+          isChanged = true;
         }
 
         if (isNumber(y)) {
           canvas.top = y;
-          changed = true;
+          isChanged = true;
         }
 
-        if (changed) {
+        if (isChanged) {
           this.renderCanvas(true);
         }
       }
@@ -174,9 +183,9 @@
      * Zoom the canvas with a relative ratio
      *
      * @param {Number} ratio
-     * @param {Event} _originalEvent (private)
+     * @param {jQuery Event} _event (private)
      */
-    zoom: function (ratio, _originalEvent) {
+    zoom: function (ratio, _event) {
       var canvas = this.canvas;
 
       ratio = num(ratio);
@@ -187,45 +196,70 @@
         ratio = 1 + ratio;
       }
 
-      this.zoomTo(canvas.width * ratio / canvas.naturalWidth, _originalEvent);
+      this.zoomTo(canvas.width * ratio / canvas.naturalWidth, _event);
     },
 
     /**
      * Zoom the canvas to an absolute ratio
      *
      * @param {Number} ratio
-     * @param {Event} _originalEvent (private)
+     * @param {jQuery Event} _event (private)
      */
-    zoomTo: function (ratio, _originalEvent) {
+    zoomTo: function (ratio, _event) {
       var options = this.options;
       var canvas = this.canvas;
       var width = canvas.width;
       var height = canvas.height;
       var naturalWidth = canvas.naturalWidth;
       var naturalHeight = canvas.naturalHeight;
+      var originalEvent;
       var newWidth;
       var newHeight;
+      var offset;
+      var center;
 
       ratio = num(ratio);
 
-      if (ratio >= 0 && this.built && !this.disabled && options.zoomable) {
+      if (ratio >= 0 && this.isBuilt && !this.isDisabled && options.zoomable) {
         newWidth = naturalWidth * ratio;
         newHeight = naturalHeight * ratio;
 
+        if (_event) {
+          originalEvent = _event.originalEvent;
+        }
+
         if (this.trigger(EVENT_ZOOM, {
-          originalEvent: _originalEvent,
+          originalEvent: originalEvent,
           oldRatio: width / naturalWidth,
           ratio: newWidth / naturalWidth
         }).isDefaultPrevented()) {
           return;
         }
 
-        canvas.left -= (newWidth - width) / 2;
-        canvas.top -= (newHeight - height) / 2;
+        if (originalEvent) {
+          offset = this.$cropper.offset();
+          center = originalEvent.touches ? getTouchesCenter(originalEvent.touches) : {
+            pageX: _event.pageX || originalEvent.pageX || 0,
+            pageY: _event.pageY || originalEvent.pageY || 0
+          };
+
+          // Zoom from the triggering point of the event
+          canvas.left -= (newWidth - width) * (
+            ((center.pageX - offset.left) - canvas.left) / width
+          );
+          canvas.top -= (newHeight - height) * (
+            ((center.pageY - offset.top) - canvas.top) / height
+          );
+        } else {
+
+          // Zoom from the center of the canvas
+          canvas.left -= (newWidth - width) / 2;
+          canvas.top -= (newHeight - height) / 2;
+        }
+
         canvas.width = newWidth;
         canvas.height = newHeight;
         this.renderCanvas(true);
-        this.setDragMode(ACTION_MOVE);
       }
     },
 
@@ -247,9 +281,9 @@
     rotateTo: function (degree) {
       degree = num(degree);
 
-      if (isNumber(degree) && this.built && !this.disabled && this.options.rotatable) {
+      if (isNumber(degree) && this.isBuilt && !this.isDisabled && this.options.rotatable) {
         this.image.rotate = degree % 360;
-        this.rotated = true;
+        this.isRotated = true;
         this.renderCanvas(true);
       }
     },
@@ -263,7 +297,7 @@
      */
     scale: function (scaleX, scaleY) {
       var image = this.image;
-      var changed = false;
+      var isChanged = false;
 
       // If "scaleY" is not present, its default value is "scaleX"
       if (isUndefined(scaleY)) {
@@ -273,18 +307,18 @@
       scaleX = num(scaleX);
       scaleY = num(scaleY);
 
-      if (this.built && !this.disabled && this.options.scalable) {
+      if (this.isBuilt && !this.isDisabled && this.options.scalable) {
         if (isNumber(scaleX)) {
           image.scaleX = scaleX;
-          changed = true;
+          isChanged = true;
         }
 
         if (isNumber(scaleY)) {
           image.scaleY = scaleY;
-          changed = true;
+          isChanged = true;
         }
 
-        if (changed) {
+        if (isChanged) {
           this.renderImage(true);
         }
       }
@@ -315,10 +349,10 @@
     /**
      * Get the cropped area position and size data (base on the original image)
      *
-     * @param {Boolean} rounded (optional)
+     * @param {Boolean} isRounded (optional)
      * @return {Object} data
      */
-    getData: function (rounded) {
+    getData: function (isRounded) {
       var options = this.options;
       var image = this.image;
       var canvas = this.canvas;
@@ -326,7 +360,7 @@
       var ratio;
       var data;
 
-      if (this.built && this.cropped) {
+      if (this.isBuilt && this.isCropped) {
         data = {
           x: cropBox.left - canvas.left,
           y: cropBox.top - canvas.top,
@@ -338,7 +372,7 @@
 
         $.each(data, function (i, n) {
           n = n / ratio;
-          data[i] = rounded ? Math.round(n) : n;
+          data[i] = isRounded ? round(n) : n;
         });
 
       } else {
@@ -372,37 +406,37 @@
       var image = this.image;
       var canvas = this.canvas;
       var cropBoxData = {};
-      var rotated;
-      var scaled;
+      var isRotated;
+      var isScaled;
       var ratio;
 
       if ($.isFunction(data)) {
         data = data.call(this.element);
       }
 
-      if (this.built && !this.disabled && $.isPlainObject(data)) {
+      if (this.isBuilt && !this.isDisabled && $.isPlainObject(data)) {
         if (options.rotatable) {
           if (isNumber(data.rotate) && data.rotate !== image.rotate) {
             image.rotate = data.rotate;
-            this.rotated = rotated = true;
+            this.isRotated = isRotated = true;
           }
         }
 
         if (options.scalable) {
           if (isNumber(data.scaleX) && data.scaleX !== image.scaleX) {
             image.scaleX = data.scaleX;
-            scaled = true;
+            isScaled = true;
           }
 
           if (isNumber(data.scaleY) && data.scaleY !== image.scaleY) {
             image.scaleY = data.scaleY;
-            scaled = true;
+            isScaled = true;
           }
         }
 
-        if (rotated) {
+        if (isRotated) {
           this.renderCanvas();
-        } else if (scaled) {
+        } else if (isScaled) {
           this.renderImage();
         }
 
@@ -434,7 +468,7 @@
      * @return {Object} data
      */
     getContainerData: function () {
-      return this.built ? this.container : {};
+      return this.isBuilt ? this.container : {};
     },
 
     /**
@@ -443,7 +477,7 @@
      * @return {Object} data
      */
     getImageData: function () {
-      return this.ready ? this.image : {};
+      return this.isLoaded ? this.image : {};
     },
 
     /**
@@ -455,7 +489,7 @@
       var canvas = this.canvas;
       var data = {};
 
-      if (this.built) {
+      if (this.isBuilt) {
         $.each([
           'left',
           'top',
@@ -484,7 +518,7 @@
         data = data.call(this.$element);
       }
 
-      if (this.built && !this.disabled && $.isPlainObject(data)) {
+      if (this.isBuilt && !this.isDisabled && $.isPlainObject(data)) {
         if (isNumber(data.left)) {
           canvas.left = data.left;
         }
@@ -514,7 +548,7 @@
       var cropBox = this.cropBox;
       var data;
 
-      if (this.built && this.cropped) {
+      if (this.isBuilt && this.isCropped) {
         data = {
           left: cropBox.left,
           top: cropBox.top,
@@ -534,14 +568,14 @@
     setCropBoxData: function (data) {
       var cropBox = this.cropBox;
       var aspectRatio = this.options.aspectRatio;
-      var widthChanged;
-      var heightChanged;
+      var isWidthChanged;
+      var isHeightChanged;
 
       if ($.isFunction(data)) {
         data = data.call(this.$element);
       }
 
-      if (this.built && this.cropped && !this.disabled && $.isPlainObject(data)) {
+      if (this.isBuilt && this.isCropped && !this.isDisabled && $.isPlainObject(data)) {
 
         if (isNumber(data.left)) {
           cropBox.left = data.left;
@@ -551,20 +585,20 @@
           cropBox.top = data.top;
         }
 
-        if (isNumber(data.width) && data.width !== cropBox.width) {
-          widthChanged = true;
+        if (isNumber(data.width)) {
+          isWidthChanged = true;
           cropBox.width = data.width;
         }
 
-        if (isNumber(data.height) && data.height !== cropBox.height) {
-          heightChanged = true;
+        if (isNumber(data.height)) {
+          isHeightChanged = true;
           cropBox.height = data.height;
         }
 
         if (aspectRatio) {
-          if (widthChanged) {
+          if (isWidthChanged) {
             cropBox.height = cropBox.width / aspectRatio;
-          } else if (heightChanged) {
+          } else if (isHeightChanged) {
             cropBox.width = cropBox.height * aspectRatio;
           }
         }
@@ -592,8 +626,12 @@
       var context;
       var data;
 
-      if (!this.built || !this.cropped || !SUPPORT_CANVAS) {
+      if (!this.isBuilt || !SUPPORT_CANVAS) {
         return;
+      }
+
+      if (!this.isCropped) {
+        return getSourceCanvas(this.$clone[0], this.image);
       }
 
       if (!$.isPlainObject(options)) {
@@ -618,8 +656,9 @@
         }
       }
 
-      canvasWidth = scaledWidth || originalWidth;
-      canvasHeight = scaledHeight || originalHeight;
+      // The canvas element will use `Math.floor` on a float number, so floor first
+      canvasWidth = floor(scaledWidth || originalWidth);
+      canvasHeight = floor(scaledHeight || originalHeight);
 
       canvas = $('<canvas>')[0];
       canvas.width = canvasWidth;
@@ -636,11 +675,12 @@
         var source = getSourceCanvas(this.$clone[0], this.image);
         var sourceWidth = source.width;
         var sourceHeight = source.height;
-        var args = [source];
+        var canvas = this.canvas;
+        var params = [source];
 
         // Source canvas
-        var srcX = data.x;
-        var srcY = data.y;
+        var srcX = data.x + canvas.naturalWidth * (abs(data.scaleX || 1) - 1) / 2;
+        var srcY = data.y + canvas.naturalHeight * (abs(data.scaleY || 1) - 1) / 2;
         var srcWidth;
         var srcHeight;
 
@@ -672,7 +712,8 @@
           srcHeight = dstHeight = min(originalHeight, sourceHeight - srcY);
         }
 
-        args.push(srcX, srcY, srcWidth, srcHeight);
+        // All the numerical parameters should be integer for `drawImage` (#476)
+        params.push(floor(srcX), floor(srcY), floor(srcWidth), floor(srcHeight));
 
         // Scale destination sizes
         if (scaledRatio) {
@@ -684,10 +725,10 @@
 
         // Avoid "IndexSizeError" in IE and Firefox
         if (dstWidth > 0 && dstHeight > 0) {
-          args.push(dstX, dstY, dstWidth, dstHeight);
+          params.push(floor(dstX), floor(dstY), floor(dstWidth), floor(dstHeight));
         }
 
-        return args;
+        return params;
       }).call(this));
 
       return canvas;
@@ -701,15 +742,15 @@
     setAspectRatio: function (aspectRatio) {
       var options = this.options;
 
-      if (!this.disabled && !isUndefined(aspectRatio)) {
+      if (!this.isDisabled && !isUndefined(aspectRatio)) {
 
         // 0 -> NaN
         options.aspectRatio = max(0, aspectRatio) || NaN;
 
-        if (this.built) {
+        if (this.isBuilt) {
           this.initCropBox();
 
-          if (this.cropped) {
+          if (this.isCropped) {
             this.renderCropBox();
           }
         }
@@ -726,8 +767,8 @@
       var croppable;
       var movable;
 
-      if (this.ready && !this.disabled) {
-        croppable = options.dragCrop && mode === ACTION_CROP;
+      if (this.isLoaded && !this.isDisabled) {
+        croppable = mode === ACTION_CROP;
         movable = options.movable && mode === ACTION_MOVE;
         mode = (croppable || movable) ? mode : ACTION_NONE;
 
@@ -746,4 +787,4 @@
         }
       }
     }
-  });
+  };