+++ src/js/preview.js
@@ -1,10 +1,11 @@
-  $.extend(prototype, {
     initPreview: function () {
       var crossOrigin = getCrossOrigin(this.crossOrigin);
-      var url = this.url;
+      var url = crossOrigin ? this.crossOriginUrl : this.url;
+      var $clone2;
 
       this.$preview = $(this.options.preview);
-      this.$viewBox.html('<img' + crossOrigin + ' src="' + url + '">');
+      this.$clone2 = $clone2 = $('<img' + crossOrigin + ' src="' + url + '">');
+      this.$viewBox.html($clone2);
       this.$preview.each(function () {
         var $this = $(this);
 
@@ -12,7 +13,7 @@
         $this.data(DATA_PREVIEW, {
           width: $this.width(),
           height: $this.height(),
-          original: $this.html()
+          html: $this.html()
         });
 
         /**
@@ -25,7 +26,7 @@
           'display:block;width:100%;height:auto;' +
           'min-width:0!important;min-height:0!important;' +
           'max-width:none!important;max-height:none!important;' +
-          'image-orientation:0deg!important">'
+          'image-orientation:0deg!important;">'
         );
       });
     },
@@ -33,8 +34,12 @@
     resetPreview: function () {
       this.$preview.each(function () {
         var $this = $(this);
+        var data = $this.data(DATA_PREVIEW);
 
-        $this.html($this.data(DATA_PREVIEW).original).removeData(DATA_PREVIEW);
+        $this.css({
+          width: data.width,
+          height: data.height
+        }).html(data.html).removeData(DATA_PREVIEW);
       });
     },
 
@@ -49,11 +54,11 @@
       var left = cropBox.left - canvas.left - image.left;
       var top = cropBox.top - canvas.top - image.top;
 
-      if (!this.cropped || this.disabled) {
+      if (!this.isCropped || this.isDisabled) {
         return;
       }
 
-      this.$viewBox.find('img').css({
+      this.$clone2.css({
         width: width,
         height: height,
         marginLeft: -left,
@@ -81,7 +86,10 @@
           newHeight = originalHeight;
         }
 
-        $this.width(newWidth).height(newHeight).find('img').css({
+        $this.css({
+          width: newWidth,
+          height: newHeight
+        }).find('img').css({
           width: width * ratio,
           height: height * ratio,
           marginLeft: -left * ratio,
@@ -89,5 +97,4 @@
           transform: getTransform(image)
         });
       });
-    }
-  });
+    },