+++ test/methods/clear.js
@@ -8,9 +8,9 @@ $(function () {
     built: function () {
       var cropper = $image.data('cropper');
 
-      QUnit.test('methods.clear', function (assert) {
+      QUnit.test('methods#clear', function (assert) {
         $image.cropper('clear');
-        assert.ok(!cropper.cropped);
+        assert.equal(cropper.isCropped, false);
         assert.ok(cropper.$cropBox.is(':hidden'));
         assert.deepEqual($image.cropper('getCropBoxData'), {});
       });