+++ test/methods/move.js
@@ -5,11 +5,9 @@ $(function () {
   var $image = $(window.createCropperImage());
 
   $image.cropper({
-    strict: false,
-
     built: function () {
 
-      QUnit.test('methods.move', function (assert) {
+      QUnit.test('methods#move', function (assert) {
         var canvasData = $image.cropper('getCanvasData');
         var changedCanvasData = $image.cropper('move', 1, 1).cropper('getCanvasData');