+++ test/methods/moveTo.js
@@ -5,11 +5,9 @@ $(function () {
   var $image = $(window.createCropperImage());
 
   $image.cropper({
-    strict: false,
-
     built: function () {
 
-      QUnit.test('methods.moveTo', function (assert) {
+      QUnit.test('methods#moveTo', function (assert) {
         var canvasData = $image.cropper('moveTo', 0, 0).cropper('getCanvasData');
 
         assert.equal(canvasData.left, 0);