+++ test/methods/reset.js
@@ -5,13 +5,11 @@ $(function () {
   var $image = $(window.createCropperImage());
 
   $image.cropper({
-    strict: false,
-
     built: function () {
       var canvasData = $image.cropper('getCanvasData');
       var cropBoxData = $image.cropper('getCropBoxData');
 
-      QUnit.test('methods.reset', function (assert) {
+      QUnit.test('methods#reset', function (assert) {
         $image.cropper('setCanvasData', {
           top: canvasData.top + 10,
           width: canvasData.width - 10