+++ test/options/autoCrop.js
@@ -7,17 +7,11 @@ $(function () {
   $image.cropper({
     autoCrop: false,
 
-    crop: function () {
-      QUnit.test('options.autoCrop', function (assert) {
-        assert.ok(false);
-      });
-    },
-
     built: function () {
       var cropper = $image.data('cropper');
 
-      QUnit.test('options.autoCrop', function (assert) {
-        assert.notEqual(cropper.cropped, true);
+      QUnit.test('options#autoCrop', function (assert) {
+        assert.notOk(cropper.isCropped);
       });
 
     }