+++ test/options/minCanvasHeight.js
@@ -6,12 +6,11 @@ $(function () {
   var minCanvasHeight = 90;
 
   $image.cropper({
-    strict: false,
     minCanvasHeight: minCanvasHeight,
 
     built: function () {
 
-      QUnit.test('options.minCanvasHeight', function (assert) {
+      QUnit.test('options#minCanvasHeight', function (assert) {
         var data = $image.cropper('setCanvasData', {
               height: 45
             }).cropper('getCanvasData');