+++ test/options/minCanvasWidth.js
@@ -6,12 +6,11 @@ $(function () {
   var minCanvasWidth = 160;
 
   $image.cropper({
-    strict: false,
     minCanvasWidth: minCanvasWidth,
 
     built: function () {
 
-      QUnit.test('options.minCanvasWidth', function (assert) {
+      QUnit.test('options#minCanvasWidth', function (assert) {
         var data = $image.cropper('setCanvasData', {
               width: 80
             }).cropper('getCanvasData');